LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-iframe-expected.txt



FAIL Starts rendering with flip-block assert_equals: expected 200 but got 0
FAIL No successful position, keep flip-block assert_equals: expected 250 but got 0
PASS Base position without fallback now successful
